Remove BikeStoreDetailActivity and its dependency. Adjust BikeStoreDetailFragment to the new logic.
