+++ app/src/main/AndroidManifest.xml
@@ -16,9 +16,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
-        <activity
-            android:name=".BikeStoreDetailActivity"
-            android:label="@string/detail" />
         <activity android:name=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
