+++ app/src/main/java/app/sram/bikestore/BikeStoreDetailFragment.kt
@@ -1,29 +1,36 @@
 package app.sram.bikestore
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.view.*
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import app.sram.bikestore.data.BikeStoreItem
+import app.sram.bikestore.databinding.FragmentBikeStoreDetailBinding
+import app.sram.bikestore.util.formatDistance
+import coil.load
 
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [BikeStoreDetailFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class BikeStoreDetailFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+
+    companion object {
+        const val ARG_STORE_ITEM = "KEY_BIKE_STORE_ITEM"
+        fun newInstance(bikeStoreItem: BikeStoreItem) =
+            BikeStoreDetailFragment().apply {
+                arguments = Bundle().apply {
+                    putParcelable(ARG_STORE_ITEM, bikeStoreItem)
+                }
+            }
+    }
+
+    private lateinit var bikeStoreItem: BikeStoreItem
+    private lateinit var binding: FragmentBikeStoreDetailBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            bikeStoreItem = it.getParcelable(ARG_STORE_ITEM)!!
         }
     }
 
@@ -31,28 +38,61 @@ class BikeStoreDetailFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_bike_store_detail, container, false)
+    ): View {
+        binding = FragmentBikeStoreDetailBinding.inflate(layoutInflater)
+        (requireActivity() as AppCompatActivity).setSupportActionBar(binding.detailActivityToolbar)
+        return binding.root
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment BikeStoreDetailFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            BikeStoreDetailFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        bindData()
+    }
+
+    private fun bindData() {
+
+        binding.storeDistTv.text = formatDistance(bikeStoreItem.distance)
+        bikeStoreItem.apply {
+            binding.storePhotoIv.load(photoUrl)
+            binding.storeNameTv.text = name
+            binding.storeAddressTv.text = vicinity
+            parentFragmentManager.beginTransaction()
+                .replace(
+                    R.id.map_fragment_view,
+                    GoogleMapWrapperFragment.newInstance(location)
+                )
+                .commit()
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.menu_share, menu)
+        super.onCreateOptionsMenu(menu, inflater)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.share -> shareToEmail(
+                "Check out ${bikeStoreItem.name} ",
+                "This cool store's address is ${bikeStoreItem.vicinity} "
+            )
+            else -> {
             }
+        }
+        return true
+    }
+
+    private fun shareToEmail(title: String, subject: String) {
+        val intent = Intent(Intent.ACTION_SENDTO).apply {
+            type = "text/html"
+            data = Uri.parse("mailto:")
+            putExtra(Intent.EXTRA_SUBJECT, title)
+            putExtra(Intent.EXTRA_TEXT, subject)
+        }
+        if (requireActivity().packageManager != null &&
+            intent.resolveActivity(requireActivity().packageManager) != null
+        ) {
+            startActivity(intent)
+        }
     }
 }
