+++ app/src/main/java/app/sram/bikestore/di/AppComponent.kt
@@ -1,7 +1,6 @@
 package app.sram.bikestore.di
 
 import app.sram.bikestore.App
-import app.sram.bikestore.BikeStoreDetailActivity
 import app.sram.bikestore.di.network.RepoModule
 import app.sram.bikestore.di.thread.SchedulerModule
 import app.sram.bikestore.di.room.RoomModule
@@ -23,9 +22,6 @@ import javax.inject.Singleton
     ]
 )
 interface AppComponent : AndroidInjector<App> {
-
-    fun inject(activity: BikeStoreDetailActivity)
-
     @Component.Builder
     interface Builder {
         @BindsInstance
